Tried to get the sox to work in VSCode. It is still not working.
