+++ src/extension.ts
@@ -1,6 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import {startListening} from './inputProcessor/speechRecognition';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -23,10 +24,15 @@ export function activate(context: vscode.ExtensionContext) {
 		// The code you place here will be executed every time your command is executed
 		// Display a message box to the user
 		vscode.window.showInformationMessage('You can start to talk');
+		//startListening();
 	});
+
+	let disposable3 = vscode.commands.registerCommand('reactalk.startListening', startListening);
+
 
 	context.subscriptions.push(disposable);
 	context.subscriptions.push(disposable2);
+	context.subscriptions.push(disposable3);
 }
 
 // this method is called when your extension is deactivated
